# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Bakery API
  description: >
    Order management for the bakery: shop заявки, baker matrix, production
    sheets (отработка), dough monitoring, and admin dictionaries. All error
    responses use the `{"error": "<safe message>"}` envelope.
  version: 1.0.0
servers:
  - url: /
    description: Direct backend access
...
security:
  - bearer: []
  - tma: []
paths:
...
  /catalog:
    get:
      tags: [orders]
      summary: Dish catalog for the order editor
      operationId: getCatalog
      responses:
        "200":
          description: Catalog items in display order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DishCatalogItem" }
        "401": { $ref: "#/components/responses/Error" }
  /categories:
    get:
      tags: [orders]
      summary: Order categories (типы заявок)
      operationId: listCategories
      responses:
        "200":
          description: Categories sorted by sort_order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
        "401": { $ref: "#/components/responses/Error" }
  /orders:
    get:
      tags: [orders]
      summary: List orders (paged, filtered)
      description: Shop users see their own shop's orders; baker/admin see all.
      operationId: listOrders
      parameters:
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 10 } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: from_department_id, in: query, schema: { type: integer, format: int64 } }
        - { name: category_id, in: query, schema: { type: integer, format: int64 } }
        - { name: fulfillment_date, in: query, schema: { type: string, format: date } }
        - { name: fulfillment_from, in: query, schema: { type: string, format: date } }
        - { name: fulfillment_to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Orders page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrdersPage" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      tags: [orders]
      summary: Create an order (shop/baker/admin; baker source is the workshop)
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderWrite" }
      responses:
        "200":
          description: Created order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /orders/draft:
    post:
      tags: [orders]
      summary: Save (upsert) the caller's draft for a category (shop only)
      description: One draft per user per category; saving again for the same category overwrites it.
      operationId: saveOrderDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderWrite" }
      responses:
        "200":
          description: Saved draft
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDraft" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /orders/drafts:
    get:
      tags: [orders]
      summary: List the caller's drafts, one per category (shop only)
      operationId: listOrderDrafts
      responses:
        "200":
          description: Drafts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrderDraft" }
        "403": { $ref: "#/components/responses/Error" }
  /orders/draft/{categoryId}:
    get:
      tags: [orders]
      summary: Get the caller's draft for a category (shop only)
      operationId: getOrderDraft
      parameters: [{ $ref: "#/components/parameters/DraftCategoryID" }]
      responses:
        "200":
          description: The draft
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDraft" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [orders]
      summary: Delete the caller's draft for a category (shop only)
      operationId: deleteOrderDraft
      parameters: [{ $ref: "#/components/parameters/DraftCategoryID" }]
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
  /orders/{id}:
    get:
      tags: [orders]
      summary: Order by number
      operationId: getOrder
      parameters: [{ $ref: "#/components/parameters/OrderNumber" }]
      responses:
        "200":
          description: The order with history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [orders]
      summary: Update an order (shop/baker/admin; category never changes)
      operationId: updateOrder
      parameters: [{ $ref: "#/components/parameters/OrderNumber" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderWrite" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /orders/{id}/cancel:
    post:
      tags: [orders]
      summary: Soft-cancel an order (idempotent)
      operationId: cancelOrder
      parameters: [{ $ref: "#/components/parameters/OrderNumber" }]
      responses:
        "200":
          description: The cancelled order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /orders/{id}/restore:
    post:
      tags: [orders]
      summary: Restore a cancelled order (idempotent)
      operationId: restoreOrder
      parameters: [{ $ref: "#/components/parameters/OrderNumber" }]
      responses:
        "200":
          description: The restored order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /production:
    post:
      tags: [production]
      summary: Create a production sheet (baker/admin)
      description: >
        Fixes the batch: every selected order and each item's loaded quantity
        are saved on the sheet; output facts become deviations from the заявка.
      operationId: createProductionSheet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductionWrite" }
      responses:
        "200":
          description: The created sheet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductionSheet" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [production]
      summary: Production journal (baker/admin)
      operationId: listProductionSheets
      responses:
        "200":
          description: Sheets, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProductionSheet" }
        "403": { $ref: "#/components/responses/Error" }
  /production/{id}:
    get:
      tags: [production]
      summary: Production sheet with items
      operationId: getProductionSheet
      parameters: [{ $ref: "#/components/parameters/SheetID" }]
      responses:
        "200":
          description: The sheet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductionSheet" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [production]
      summary: Replace a sheet's batch and deviations
      description: >
        Output values equal to the order quantity drop the deviation row;
        loaded quantities remain stored. The sheet keeps its batch and lives
        until explicitly deleted, even with no output deviations left.
      operationId: updateProductionSheet
      parameters: [{ $ref: "#/components/parameters/SheetID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductionWrite" }
      responses:
        "200":
          description: Updated sheet, or `id:0` when deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductionSheet" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [production]
      summary: Delete a sheet (facts fall back to the заявка)
      operationId: deleteProductionSheet
      parameters: [{ $ref: "#/components/parameters/SheetID" }]
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: HMAC web session token from POST /login (7-day TTL).
    tma:
      type: apiKey
      in: header
      name: Authorization
      description: '`tma <initData>` — Telegram Mini App initData, validated per request (24h max age).'
  parameters:
    OrderNumber:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Order number, e.g. Г.Х.08.07.26.001
    SheetID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    DraftCategoryID:
      name: categoryId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    Error:
      description: Error with a safe user-facing message
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
    Department:
      type: object
      required: [id, code, name, type]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: gagarina }
        name: { type: string, example: Магазин Гагарина }
        type: { type: string, enum: [shop, workshop] }
...
    Category:
      type: object
      required: [id, code, letter, name, color, sort_order, monitor_codes]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: bread }
        letter: { type: string, example: Х, description: Becomes part of the order number }
        name: { type: string, example: Хлеб }
        color:
          type: string
          enum: [amber, sky, violet, emerald, rose, stone]
          description: Must stay in sync with domain.CategoryColors and frontend lib/categories.js
        sort_order: { type: integer, format: int64 }
        monitor_codes:
          type: array
          items: { type: string }
          description: Dough codes used by /monitor for this category
...
    OrderItem:
      type: object
      required: [code, product_name, quantity, reserved_quantity, production_quantity]
      properties:
        code: { type: string }
        product_name: { type: string }
        quantity: { type: number }
        reserved_quantity: { type: number }
        production_quantity: { type: number, description: quantity + reserved_quantity }
        produced_quantity:
          type: number
          nullable: true
          description: Production fact; absent = baked as ordered
        produced_reason: { type: string }
    Comments:
      type: object
      required: [general, items]
      properties:
        general: { type: string }
        items:
          type: array
          items:
            type: object
            required: [product_name, comment]
            properties:
              product_name: { type: string }
              comment: { type: string }
    HistoryItem:
      type: object
      required: [change_type, product_code, product_name]
      properties:
        change_type: { type: string, enum: [added, updated, removed] }
        product_code: { type: string }
        product_name: { type: string }
        old_quantity: { type: number, nullable: true }
        new_quantity: { type: number, nullable: true }
        old_reserved_quantity: { type: number, nullable: true }
        new_reserved_quantity: { type: number, nullable: true }
    HistoryEntry:
      type: object
      required: [id, changed_by_username, changed_at, items]
      properties:
        id: { type: integer, format: int64 }
        changed_by_username: { type: string }
        changed_at: { type: string, description: RFC3339 }
        items:
          type: array
          items: { $ref: "#/components/schemas/HistoryItem" }
    Order:
      type: object
      required: [id, number, location, created_by_username, items, created_at, fulfillment_date, monitor_command, comments, favorite, cancelled]
      properties:
        id: { type: string }
        number: { type: string, example: Г.Х.08.07.26.001 }
        location: { type: string }
        created_by_username: { type: string }
        from_department: { $ref: "#/components/schemas/Department" }
        to_department: { $ref: "#/components/schemas/Department" }
        category: { $ref: "#/components/schemas/Category" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        created_at: { type: string, description: RFC3339 }
        fulfillment_date: { type: string, format: date }
        monitor_command: { type: string, example: /monitor Г.Х.08.07.26.001 }
        comments: { $ref: "#/components/schemas/Comments" }
        favorite: { type: boolean }
        cancelled: { type: boolean }
        cancelled_by_username: { type: string }
        production_sheet_id:
          type: integer
          format: int64
          nullable: true
          description: The sheet covering this order (an order has at most one)
        history:
          type: array
          items: { $ref: "#/components/schemas/HistoryEntry" }
    OrdersPage:
      type: object
      required: [items, page, limit, offset, total, total_pages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Order" }
        page: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer, format: int64 }
        total_pages: { type: integer }
    OrderWrite:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [product_name, quantity]
            properties:
              product_name: { type: string }
              quantity: { type: number }
              reserved_quantity: { type: number }
        fulfillment_date: { type: string, format: date }
        from_department_id: { type: integer, format: int64, nullable: true }
        category_id:
          type: integer
          format: int64
          description: Required on create; ignored on update (category is immutable)
        comments: { $ref: "#/components/schemas/Comments" }
    OrderDraft:
      type: object
      required: [write, updated_at]
      properties:
        write: { $ref: "#/components/schemas/OrderWrite" }
        updated_at: { type: string, format: date-time }
    DishCatalogItem:
      type: object
      required: [code, name, theme, sort_order]
      properties:
        code: { type: string }
        name: { type: string }
        theme: { type: string, description: "Group name, e.g. КОКРОКИ (normalized to a dish_groups row keyed by category+name)" }
        category_id: { type: integer, format: int64, nullable: true }
        sort_order: { type: integer, format: int64 }
...
    ProductionItemWrite:
      type: object
      required: [product_name, loaded_quantity, produced_quantity]
      properties:
        product_name: { type: string }
        loaded_quantity: { type: number, minimum: 0, description: Закладка — количество, переданное в производство }
        produced_quantity: { type: number, minimum: 0, description: Выход — фактически полученное количество }
        reason: { type: string, maxLength: 200, description: Optional justification («подгорело», «упало») }
    ProductionWrite:
      type: object
      required: [orders]
      properties:
        orders:
          description: The batch — every selected order, deviations optional per order
          type: array
          items:
            type: object
            required: [number, items]
            properties:
              number: { type: string }
              items:
                type: array
                items: { $ref: "#/components/schemas/ProductionItemWrite" }
    ProductionSheetItem:
      type: object
      required: [order_number, product_name, loaded_quantity, produced_quantity, reason]
      properties:
        order_number: { type: string }
        product_name: { type: string }
        loaded_quantity: { type: number }
        produced_quantity: { type: number }
        reason: { type: string }
    ProductionSheet:
      type: object
      required: [id, created_by_username, created_at, updated_at, order_numbers, item_count]
      properties:
        id: { type: integer, format: int64 }
        created_by_username: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
        order_numbers:
          type: array
          items: { type: string }
          description: The saved batch — every covered order, including those without deviations
        item_count: { type: integer, format: int64, description: Number of deviation rows }
        items:
          type: array
          items: { $ref: "#/components/schemas/ProductionSheetItem" }
...