# internal/services/order/infra/http/handler.go
package orderhttp

import (
	"context"
	"fmt"
	"log/slog"
	"math"
	"net/http"
	"strconv"

	"bakery/internal/inbound/api/contract"
	"bakery/internal/inbound/api/httpx"
	departmentuc "bakery/internal/services/department/usecase/department"
	orderdomain "bakery/internal/services/order/domain"
	orderuc "bakery/internal/services/order/usecase/order"
	techcarduc "bakery/internal/services/techcard/usecase/techcard"
)

// Handler is the order service HTTP delivery adapter.
type Handler struct {
	orderSvc      orderuc.UseCase
	departmentSvc departmentuc.UseCase
	techcardSvc   techcarduc.UseCase
	presenter     *OrderPresenter
}

func New(orderSvc orderuc.UseCase, departmentSvc departmentuc.UseCase, techcardSvc techcarduc.UseCase, presenter *OrderPresenter) *Handler {
	return &Handler{orderSvc: orderSvc, departmentSvc: departmentSvc, techcardSvc: techcardSvc, presenter: presenter}
}

// RegisterRoutes wires the order endpoints behind Mini App authentication.
func (h *Handler) RegisterRoutes(mux *http.ServeMux, auth *httpx.Authenticator) {
	mux.Handle("GET /catalog", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleCatalog)))
	mux.Handle("GET /categories", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleListCategories)))
	mux.Handle("GET /orders", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleListOrders)))
	mux.Handle("POST /orders", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleCreateOrder)))
	mux.Handle("GET /orders/{id}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleOrderByID)))
	mux.Handle("PUT /orders/{id}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleUpdateOrder)))
	mux.Handle("POST /orders/{id}/cancel", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleCancelOrder)))
	mux.Handle("POST /orders/{id}/restore", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleRestoreOrder)))
	mux.Handle("POST /orders/draft", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleSaveOrderDraft)))
	mux.Handle("GET /orders/drafts", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleListOrderDrafts)))
	mux.Handle("GET /orders/draft/{categoryId}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleGetOrderDraft)))
	mux.Handle("DELETE /orders/draft/{categoryId}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleDeleteOrderDraft)))
	mux.Handle("POST /production", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleCreateProductionSheet)))
	mux.Handle("GET /production", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleListProductionSheets)))
	mux.Handle("GET /production/{id}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleGetProductionSheet)))
	mux.Handle("PUT /production/{id}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleUpdateProductionSheet)))
	mux.Handle("DELETE /production/{id}", auth.RequireMiniAppAuth(http.HandlerFunc(h.handleDeleteProductionSheet)))
	h.RegisterAdminRoutes(mux, auth)
}

func (h *Handler) handleListOrders(w http.ResponseWriter, r *http.Request) {
	if h.orderSvc == nil {
		httpx.WriteError(w, http.StatusServiceUnavailable, "Сервис заказов временно недоступен.")
		return
	}

	limit := int32(10)
	if raw := httpx.Trim(r.URL.Query().Get("limit")); raw != "" {
		parsed, err := strconv.ParseInt(raw, 10, 32)
		if err != nil || parsed <= 0 {
			httpx.WriteError(w, http.StatusBadRequest, "Параметр limit должен быть положительным числом.")
			return
		}
		limit = int32(parsed)
	}
	page := int32(1)
	if raw := httpx.Trim(r.URL.Query().Get("page")); raw != "" {
		parsed, err := strconv.ParseInt(raw, 10, 32)
		if err != nil || parsed <= 0 {
			httpx.WriteError(w, http.StatusBadRequest, "Параметр page должен быть положительным числом.")
			return
		}
		page = int32(parsed)
	}
	offset := (page - 1) * limit
	var fromDepartmentID *int64
	if raw := httpx.Trim(r.URL.Query().Get("from_department_id")); raw != "" {
		parsed, err := strconv.ParseInt(raw, 10, 64)
		if err != nil || parsed <= 0 {
			httpx.WriteError(w, http.StatusBadRequest, "Параметр from_department_id должен быть положительным числом.")
			return
		}
		fromDepartmentID = &parsed
	}
	var categoryID *int64
	if raw := httpx.Trim(r.URL.Query().Get("category_id")); raw != "" {
		parsed, err := strconv.ParseInt(raw, 10, 64)
		if err != nil || parsed <= 0 {
			httpx.WriteError(w, http.StatusBadRequest, "Параметр category_id должен быть положительным числом.")
			return
		}
		categoryID = &parsed
	}
	fulfillmentDate, err := httpx.ParseRequestDate(r.URL.Query().Get("fulfillment_date"))
	if err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "Дата выполнения должна быть в формате YYYY-MM-DD.")
		return
	}
	fulfillmentFrom, err := httpx.ParseRequestDate(r.URL.Query().Get("fulfillment_from"))
	if err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "Дата fulfillment_from должна быть в формате YYYY-MM-DD.")
		return
	}
	fulfillmentTo, err := httpx.ParseRequestDate(r.URL.Query().Get("fulfillment_to"))
	if err != nil {
		httpx.WriteError(w, http.StatusBadRequest, "Дата fulfillment_to должна быть в формате YYYY-MM-DD.")
		return
	}

	result, err := h.orderSvc.ListOrders(r.Context(), orderuc.ListOrdersInput{
		Limit:            limit,
		Offset:           offset,
		FromDepartmentID: fromDepartmentID,
		CategoryID:       categoryID,
		FulfillmentDate:  fulfillmentDate,
		FulfillmentFrom:  fulfillmentFrom,
		FulfillmentTo:    fulfillmentTo,
	})
	if err != nil {
		slog.ErrorContext(r.Context(), "list orders failed", "error", err)
		httpx.WriteError(w, http.StatusInternalServerError, "Не удалось получить список заказов.")
		return
	}

	httpx.WriteJSON(w, http.StatusOK, contract.OrdersPage{
		Items:      h.presenter.BuildOrderResponses(r.Context(), result.Orders),
		Page:       page,
		Limit:      result.Limit,
		Offset:     result.Offset,
		Total:      result.Total,
		TotalPages: int32(math.Ceil(float64(result.Total) / float64(result.Limit))),
	})
}

// handleListCategories returns the типы заявок for pickers, filters and cards.
func (h *Handler) handleListCategories(w http.ResponseWriter, r *http.Request) {
	if h.orderSvc == nil {
		httpx.WriteError(w, http.StatusServiceUnavailable, "Сервис заказов временно недоступен.")
		return
	}
	categories, err := h.orderSvc.ListOrderCategories(r.Context())
	if err != nil {
		slog.ErrorContext(r.Context(), "list order categories failed", "error", err)
		httpx.WriteError(w, http.StatusInternalServerError, "Не удалось загрузить типы заявок.")
		return
	}
	responses := make([]contract.Category, 0, len(categories))
	for _, category := range categories {
		responses = append(responses, *buildCategoryResponse(&category))
	}
	httpx.WriteJSON(w, http.StatusOK, responses)
}

func (h *Handler) handleOrderByID(w http.ResponseWriter, r *http.Request) {
	if h.orderSvc == nil {
		httpx.WriteError(w, http.StatusServiceUnavailable, "Сервис заказов временно недоступен.")
		return
	}

	id := httpx.Trim(r.PathValue("id"))
	if id == "" {
		httpx.WriteError(w, http.StatusBadRequest, "Укажите номер заказа.")
		return
	}

	order, err := h.orderSvc.GetOrderByNumber(r.Context(), id)
	if err != nil {
		httpx.WriteError(w, http.StatusNotFound, fmt.Sprintf("Заказ %s не найден.", id))
		return
	}
	if !UserCanReadOrder(r.Context(), order) {
		httpx.WriteError(w, http.StatusNotFound, fmt.Sprintf("Заказ %s не найден.", id))
		return
	}

	httpx.WriteJSON(w, http.StatusOK, h.presenter.BuildOrderResponse(r.Context(), order))
}

// UserCanReadOrder reports whether the request's viewer may read the order.
// Users are no longer bound to a department, so any authenticated app user
// (shop, baker, admin) may read any order.
func UserCanReadOrder(ctx context.Context, order orderdomain.Order) bool {
	_, ok := httpx.MiniAppUserFromContext(ctx)
	return ok
}